placed tests in another spec component as the ribbon does not exist in the tested component anymore
