+++ src/main/webapp/app/exam/participate/exam-participation.component.html
@@ -1,4 +1,4 @@
-<jhi-test-run-ribbon *ngIf="!!testRunId"></jhi-test-run-ribbon>
+<jhi-test-run-ribbon id="testRunRibbon" *ngIf="!!testRunId"></jhi-test-run-ribbon>
 <ng-container *ngIf="exam">
     <jhi-exam-participation-cover
         *ngIf="isVisible() && !isGracePeriodOver() && !this.studentExam?.submitted && !examStartConfirmed"
