+++ src/test/javascript/spec/component/exam/participate/exam-participation.component.spec.ts
@@ -45,6 +45,7 @@ import * as sinonChai from 'sinon-chai';
 import { TranslatePipeMock } from '../../../helpers/mocks/service/mock-translate.service';
 import { ArtemisTestModule } from '../../../test.module';
 import { FileUploadExamSubmissionComponent } from 'app/exam/participate/exercises/file-upload/file-upload-exam-submission.component';
+import { By } from '@angular/platform-browser';
 
 chai.use(sinonChai);
 const expect = chai.expect;
@@ -121,6 +122,25 @@ describe('ExamParticipationComponent', () => {
     it('should initialize', () => {
         fixture.detectChanges();
         expect(ExamParticipationComponent).to.be.ok;
+    });
+
+    describe('ExamParticipationSummaryComponent for TestRuns', () => {
+        it('should initialize and display test run ribbon', function () {
+            fixture.detectChanges();
+            expect(fixture).to.be.ok;
+            expect(!!comp.testRunId).to.be.true;
+            const testRunRibbon = fixture.debugElement.query(By.css('#testRunRibbon'));
+            expect(testRunRibbon).to.exist;
+        });
+        it('should initialize and not display test run ribbon', function () {
+            TestBed.get(ActivatedRoute).params = of({ courseId: '1', examId: '2' });
+            comp.ngOnInit();
+            fixture.detectChanges();
+            expect(fixture).to.be.ok;
+            expect(!!comp.testRunId).to.be.false;
+            const testRunRibbon = fixture.debugElement.query(By.css('#testRunRibbon'));
+            expect(testRunRibbon).to.not.exist;
+        });
     });
 
     describe('isProgrammingExercise', () => {
