+++ src/test/javascript/spec/component/exam/participate/summary/exam-participation-summary.component.spec.ts
@@ -133,28 +133,8 @@ function sharedSetup(url: string[]) {
     });
 }
 
-describe('ExamParticipationSummaryComponent for TestRuns', () => {
-    sharedSetup(['', 'test-runs']);
-
-    it('should initialize and display test run ribbon', function () {
-        fixture.detectChanges();
-        expect(fixture).to.be.ok;
-        expect(component.isTestRun).to.be.true;
-        const testRunRibbon = fixture.debugElement.query(By.css('#testRunRibbon'));
-        expect(testRunRibbon).to.exist;
-    });
-});
-
 describe('ExamParticipationSummaryComponent', () => {
     sharedSetup(['', '']);
-
-    it('should initialize and not display test run ribbon', function () {
-        fixture.detectChanges();
-        expect(fixture).to.be.ok;
-        expect(component.isTestRun).to.be.false;
-        const testRunRibbon = fixture.debugElement.query(By.css('#testRunRibbon'));
-        expect(testRunRibbon).to.not.exist;
-    });
 
     it('should expand all exercises and call print when Export PDF is clicked', fakeAsync(() => {
         const printWindowStub = sinon.stub(global.window, 'print').returns();
